delete & edit button handlers operational. edit form completed
