+++ src/app/charities/page.tsx
@@ -69,6 +69,7 @@ const Explore = () => {
                 {userProfile?.is_staff && (
                     <button 
                         className="px-4 py-2 border rounded-lg bg-white dark:bg-gray-800 dark:text-gray-200 border-gray-300 dark:border-gray-600 hover:bg-gray-100 dark:hover:bg-gray-700 focus:outline-none focus:ring-2 focus:ring-blue-500 dark:focus:ring-blue-400 flex items-center gap-2"
+                        onClick={()=>{router.push("/charities/create")}}
                         >
                         <span className="text-lg">+</span>
                         Add Organization
